Rename 'model' modules to 'settings'
