+++ rust/agama-lib/src/network.rs
@@ -1,9 +1,9 @@
 mod client;
-mod model;
+mod settings;
 mod proxies;
 mod store;
 pub mod types;
 
 pub use client::NetworkClient;
-pub use model::NetworkSettings;
+pub use settings::NetworkSettings;
 pub use store::NetworkStore;
