+++ rust/agama-lib/src/network/settings.rs
