+++ rust/agama-lib/src/software/settings.rs
