+++ rust/agama-lib/src/storage/settings.rs
