+++ rust/agama-lib/src/users.rs
@@ -1,8 +1,8 @@
 mod client;
-mod model;
+mod settings;
 mod proxies;
 mod store;
 
 pub use client::{UsersClient, FirstUser};
-pub use model::{FirstUserSettings, RootUserSettings, UserSettings};
+pub use settings::{FirstUserSettings, RootUserSettings, UserSettings};
 pub use store::UsersStore;
