+++ rust/agama-lib/src/users/settings.rs
